backend/requirements.txt
python-dotenv==0.21.1
fastapi==0.89.1
databases==0.7.0
sqlalchemy==1.4.46
uvicorn==0.20.0
psycopg2==2.9.5
pydantic[email]==1.10.4
alembic==1.9.2
asyncpg==0.27.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
pytest==7.2.1
pytest-mock==3.10.0
requests==2.28.2
bcrypt==4.0.1
httpx==0.23.3
redis==4.4.2
types-redis==4.4.0.3
aiofiles==22.1.0
types-aiofiles==22.1.0.6
Pillow==9.4.0